backend/requirements.txt
# BudgetApp - Requirements
# Python 3.12+

# FastAPI Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Data Processing
pandas==2.1.3
openpyxl==3.1.2

# CORS and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
authlib==1.3.0
httpx==0.25.2  # Required by authlib for OAuth
google-auth==2.43.0  # Google OAuth token verification
requests==2.31.0  # Required by google-auth

# Date and Time
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
